Fix SQMtools tpm calculation and PFAM name parsing
This fixes several bugs
- ORF TPM was not being rescaled when rescale_tpm was TRUE. Subsequent
subsets of the resulting object used the non-rescaled ORF TPM, leading
to inconsistencies between the parent and the subset (see #212)
- aggregate.fun was not taking into account the unmapped reads when
calculating TPMs, but sqm2tables.py is, since v1.3. This led to
inconsistencies between the TPMs of the original object and the TPMs of
its subsets. They are still present (due to the quick&dirty way
aggregate.fun calculates TPMs) but they are much smaller now.
- We now use abundances instead of bases for TPM calculation in all cases
- PFAM names containing semicolons were being corrupted by aggregate.fun